thread: move stack allocation from platform specific
Moves static stack allocation from platform specific code to the thread
subsystem.

Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (QEMU)
Tested-by: Pascal Brand <pascal.brand@linaro.org> (STM platform)
