+++ core/arch/arm32/plat-stm/conf.mk
@@ -6,7 +6,7 @@ include mk/gcc.mk
 
 core-platform-cppflags	 = -I$(arch-dir)/include
 core-platform-cppflags	+= -DNUM_THREADS=2
-core-platform-cppflags	+= -DWITH_STACK_CANARIES=1
+
 
 core-platform-subdirs += \
 	$(addprefix $(arch-dir)/, kernel mm sm tee sta) $(platform-dir)
@@ -15,6 +15,8 @@ libutil_with_isoc := y
 WITH_PL310 := y
 WITH_SECURE_TIME_SOURCE_REE := y
 CFG_CACHE_API := y
+CFG_WITH_STACK_CANARIES := y
+CFG_WITH_SEC_MON := y
 
 include mk/config.mk
 include $(platform-dir)/system_config.in
