+++ core/arch/arm32/plat-sunxi/main.c
@@ -32,13 +32,11 @@
 #include <assert.h>
 
 #include <sm/sm.h>
-#include <sm/sm_defs.h>
 #include <sm/tee_mon.h>
 #include <sm/teesmc.h>
 #include <sm/teesmc_optee.h>
 
 #include <arm32.h>
-#include <kernel/arch_debug.h>
 #include <kernel/thread.h>
 #include <kernel/time_source.h>
 #include <kernel/panic.h>
@@ -53,62 +51,6 @@
 #include <util.h>
 #include <trace.h>
 #include <malloc.h>
-
-#ifdef WITH_STACK_CANARIES
-#define STACK_CANARY_SIZE	(4 * sizeof(uint32_t))
-#define START_CANARY_VALUE	0xdededede
-#define END_CANARY_VALUE	0xabababab
-#define GET_START_CANARY(name, stack_num) name[stack_num][0]
-#define GET_END_CANARY(name, stack_num) \
-	name[stack_num][sizeof(name[stack_num]) / sizeof(uint32_t) - 1]
-#else
-#define STACK_CANARY_SIZE	0
-#endif
-
-#define STACK_ALIGNMENT		8
-
-#define DECLARE_STACK(name, num_stacks, stack_size) \
-	static uint32_t name[num_stacks][(stack_size + STACK_CANARY_SIZE) / \
-					 sizeof(uint32_t)] \
-		__attribute__((section(".bss.prebss.stack"), \
-			       aligned(STACK_ALIGNMENT)))
-
-#define GET_STACK(stack) \
-	((vaddr_t)(stack) + sizeof(stack) - STACK_CANARY_SIZE / 2)
-
-
-DECLARE_STACK(stack_tmp,	CFG_TEE_CORE_NB_CORE,	STACK_TMP_SIZE);
-DECLARE_STACK(stack_abt,	CFG_TEE_CORE_NB_CORE,	STACK_ABT_SIZE);
-DECLARE_STACK(stack_sm,		CFG_TEE_CORE_NB_CORE,	SM_STACK_SIZE);
-DECLARE_STACK(stack_thread,	NUM_THREADS,		STACK_THREAD_SIZE);
-
-const vaddr_t stack_tmp_top[CFG_TEE_CORE_NB_CORE] = {
-	GET_STACK(stack_tmp[0]),
-#if CFG_TEE_CORE_NB_CORE > 1
-	GET_STACK(stack_tmp[1]),
-#endif
-#if CFG_TEE_CORE_NB_CORE > 2
-	GET_STACK(stack_tmp[2]),
-#endif
-#if CFG_TEE_CORE_NB_CORE > 3
-	GET_STACK(stack_tmp[3]),
-#endif
-#if CFG_TEE_CORE_NB_CORE > 4
-	GET_STACK(stack_tmp[4]),
-#endif
-#if CFG_TEE_CORE_NB_CORE > 5
-	GET_STACK(stack_tmp[5]),
-#endif
-#if CFG_TEE_CORE_NB_CORE > 6
-	GET_STACK(stack_tmp[6]),
-#endif
-#if CFG_TEE_CORE_NB_CORE > 7
-	GET_STACK(stack_tmp[7]),
-#endif
-#if CFG_TEE_CORE_NB_CORE > 8
-#error "Top of tmp stacks aren't defined for more than 8 CPUS"
-#endif
-};
 
 /* teecore heap address/size is defined in scatter file */
 extern unsigned char teecore_heap_start;
@@ -132,42 +74,6 @@ static uint32_t main_mmu_ul1_ttb[NUM_THREADS][TEE_MMU_UL1_NUM_ENTRIES]
 static void main_fiq(void);
 static void main_tee_entry(struct thread_smc_args *args);
 static uint32_t main_default_pm_handler(uint32_t a0, uint32_t a1);
-
-static void init_canaries(void)
-{
-	size_t n;
-#define INIT_CANARY(name)						\
-	for (n = 0; n < ARRAY_SIZE(name); n++) {			\
-		uint32_t *start_canary = &GET_START_CANARY(name, n);	\
-		uint32_t *end_canary = &GET_END_CANARY(name, n);	\
-									\
-		*start_canary = START_CANARY_VALUE;			\
-		*end_canary = END_CANARY_VALUE;				\
-	}
-
-	INIT_CANARY(stack_tmp);
-	INIT_CANARY(stack_abt);
-	INIT_CANARY(stack_sm);
-	INIT_CANARY(stack_thread);
-}
-
-void check_canaries(void)
-{
-#ifdef WITH_STACK_CANARIES
-	size_t n;
-
-#define ASSERT_STACK_CANARIES(name)					\
-	for (n = 0; n < ARRAY_SIZE(name); n++) {			\
-		assert(GET_START_CANARY(name, n) == START_CANARY_VALUE);\
-		assert(GET_END_CANARY(name, n) == END_CANARY_VALUE);	\
-	} while (0)
-
-	ASSERT_STACK_CANARIES(stack_tmp);
-	ASSERT_STACK_CANARIES(stack_abt);
-	ASSERT_STACK_CANARIES(stack_sm);
-	ASSERT_STACK_CANARIES(stack_thread);
-#endif /*WITH_STACK_CANARIES*/
-}
 
 static const struct thread_handlers handlers = {
 	.std_smc = main_tee_entry,
@@ -198,35 +104,18 @@ void main_init(uint32_t nsec_entry)
 	write_cpsr(read_cpsr() | CPSR_F | CPSR_I);
 
 	if (pos == 0) {
-		size_t n;
-
-		/* Initialize canries around the stacks */
-		init_canaries();
-
-		/* Assign the thread stacks */
-		for (n = 0; n < NUM_THREADS; n++) {
-			if (!thread_init_stack(n, GET_STACK(stack_thread[n])))
-				panic();
-		}
-		thread_init_handlers(&handlers);
+		thread_init_primary(&handlers);
 
 		/* initialize platform */
 		platform_init();
 	}
 
-	if (!thread_init_stack(THREAD_TMP_STACK, GET_STACK(stack_tmp[pos])))
-		panic();
-	if (!thread_init_stack(THREAD_ABT_STACK, GET_STACK(stack_abt[pos])))
-		panic();
-
 	thread_init_per_cpu();
 
 	/* Initialize secure monitor */
-	sm_init(GET_STACK(stack_sm[pos]));
 	nsec_ctx = sm_get_nsec_ctx();
 	nsec_ctx->mon_lr = nsec_entry;
 	nsec_ctx->mon_spsr = CPSR_MODE_SVC | CPSR_I;
-	sm_set_entry_vector(thread_vector_table);
 
 	if (pos == 0) {
 		unsigned long a, s;
