+++ core/arch/arm32/plat-vexpress/platform_config.h
@@ -97,13 +97,6 @@
 #error "Unknown platform flavor"
 #endif
 
-#define STACK_TMP_SIZE		1024
-#if CFG_TRACE_LEVEL > 0
-#define STACK_ABT_SIZE		2048
-#else
-#define STACK_ABT_SIZE		1024
-#endif
-#define STACK_THREAD_SIZE	8192
 #define HEAP_SIZE		(24 * 1024)
 
 #if PLATFORM_FLAVOR_IS(fvp)
